Multiple definitions fix
